+++ R/MANOVA-function.R
@@ -9,7 +9,7 @@
 #'   contains the response variable and the right hand side contains the factor 
 #'   variables of interest. An interaction term must be specified.
 #' @param data A data.frame, list or environment containing the variables in 
-#'   \code{formula}. Data must be in long format.
+#'   \code{formula}. Data must be in long format and must not contain missing values.
 #' @param subject The column name of the subjects in the data.
 #' @param iter The number of iterations used for calculating the resampled 
 #'   statistic. The default option is 10,000.
@@ -29,7 +29,7 @@
 #'   
 #' @details The MANOVA() function provides the Wald-type statistic (WTS) as well as
 #'   the modified ANOVA-type statistic (MATS) for multivariate designs with metric data as described in 
-#'   Konietschke et al. (2015) and Friedrich and Pauly (2017), respectively. The MATS is invariant
+#'   Konietschke et al. (2015) and Friedrich and Pauly (2018), respectively. The MATS is invariant
 #'   under scale transformations of the components and applicable to designs with singular covariance
 #'   matrices.
 #'   Both tests are applicable for non-normal error terms, 
@@ -68,20 +68,21 @@
 #'   Friedrich, S., Brunner, E. and Pauly, M. (2017). Permuting longitudinal data
 #'   in spite of the dependencies. Journal of Multivariate Analysis, 153, 255-265.
 #'   
-#'    Bathke, A., Friedrich, S., Konietschke, F., Pauly, M., Staffen, W., Strobl, N. and Hoeller, Y. (2016).
-#'    Using EEG, SPECT, and Multivariate Resampling Methods
-#'    to Differentiate Between Alzheimer's and other Cognitive Impairments. arXiv preprint arXiv:1606.09004.
+#'   Bathke, A., Friedrich, S., Konietschke, F., Pauly, M., Staffen, W., Strobl, N. and 
+#'   Hoeller, Y. (2018). Testing Mean Differences among Groups: Multivariate and Repeated 
+#'   Measures Analysis with Minimal Assumptions. Multivariate Behavioral Research, 53(3), 348-359, 
+#'   Doi: 10.1080/00273171.2018.1446320.
 #'   
 #'   Friedrich, S., Konietschke, F., Pauly, M. (2017). GFD - An 
 #'   R-package for the Analysis of General Factorial Designs. 
 #'   Journal of Statistical Software, 79(1), 1-18.
 #'   
-#'   Friedrich, S., and Pauly, M. (2017). MATS: Inference for potentially singular and
-#'   heteroscedastic MANOVA. arXiv preprint arXiv:1704.03731.
+#'   Friedrich, S., and Pauly, M. (2018). MATS: Inference for potentially singular and
+#'   heteroscedastic MANOVA. Journal of Multivariate Analysis, 165, 166-179.
 #'    
 #'   
 #' @importFrom graphics axis legend par plot title abline points
-#' @importFrom stats ecdf formula model.frame pchisq pf qt terms var cov rbinom quantile
+#' @importFrom stats ecdf formula model.frame pchisq pf qt terms var cov rbinom quantile as.formula
 #' @importFrom utils read.table
 #' @importFrom methods hasArg
 #' @importFrom MASS mvrnorm
@@ -96,6 +97,10 @@ MANOVA <- function(formula, data, subject,
   
   if (!(resampling %in% c("paramBS", "WildBS"))){
     stop("Resampling must be one of 'paramBS' and 'WildBS'!")
+  }
+  
+  if(sum(grepl("cbind", formula)) != 0){
+    stop("For data in wide format, please use function MANOVA.wide()")
   }
   
   input_list <- list(formula = formula, data = data,
@@ -117,6 +122,10 @@ MANOVA <- function(formula, data, subject,
     stop("The subject variable is not found!")
   }
   subject <- data[, subject]
+  if (length(subject) != nrow(dat)){
+    stop("There are missing values in the data.")
+  }
+  
   # no. of dimensions 
   p <- length(subject)/length(unique(subject))
   dat2 <- data.frame(dat, subject = subject)
@@ -132,14 +141,23 @@ MANOVA <- function(formula, data, subject,
     levels[[jj]] <- levels(as.factor(dat[, (jj + 1)]))
   }
   lev_names <- expand.grid(levels)
+  
+  # number of hypotheses
+  tmp <- 0
+  for (i in 1:nf) {
+    tmp <- c(tmp, choose(nf, i))
+    nh <- sum(tmp)
+  }
+  
   if (nf == 1) {
     # one-way layout
+    nest <- FALSE
     nr_hypo <- attr(terms(formula), "factors")
     fac_names <- colnames(nr_hypo)
     dat2 <- dat2[order(dat2[, 2], dat2$subject), ]
     response <- dat2[, 1]    
     # contrast matrix
-    hypo <- (diag(fl) - matrix(1 / fl, ncol = fl, nrow = fl)) %x% diag(p)
+    hypo_matrices <- (diag(fl) - matrix(1 / fl, ncol = fl, nrow = fl)) %x% diag(p)
     n <- plyr::ddply(dat2, nadat2, plyr::summarise, Measure = length(unique(subject)),
                      .drop = F)$Measure
     WTS_out <- matrix(NA, ncol = 3, nrow = 1)
@@ -148,7 +166,7 @@ MANOVA <- function(formula, data, subject,
     quantiles <- matrix(NA, 2, 1)
     rownames(WTS_out) <- fac_names
     names(WTPS_out) <- fac_names
-    results <- MANOVA.Stat(data = response, n = n, hypo, iter = iter, alpha, resampling, n.groups = fl, p, CPU, seed, nf)    
+    results <- MANOVA.Stat(data = response, n = n, hypo_matrices, iter = iter, alpha, resampling, n.groups = fl, p, CPU, seed, nf)    
     WTS_out <- round(results$WTS, dec)
     MATS_out <- round(results$MATS, dec)
     WTPS_out <- round(results$WTPS, dec)
@@ -161,26 +179,14 @@ MANOVA <- function(formula, data, subject,
     names(WTS_out) <- cbind ("Test statistic", "df",
                              "p-value")
     names(WTPS_out) <- cbind(paste(resampling, "(WTS)"), paste(resampling, "(MATS)"))
-    output <- list()
-    output$input <- input_list
-    output$Descriptive <- descriptive
-    output$Covariance <- Var_out
-    output$WTS <- WTS_out
-    output$MATS <- MATS_out
-    output$resampling <- WTPS_out
-    output$quantile <- quantiles
-    output$nf <- nf
-    output$factors <- fac_names
-    output$H <- hypo
-    output$p <- p
-    output$fl <- fl
-    output$Means <- mean_out
+    colnames(MATS_out) <- "Test statistic"
     # end one-way layout ------------------------------------------------------
   } else {
     dat2 <- dat2[do.call(order, dat2[, 2:(nf + 2)]), ]
     lev_names <- lev_names[do.call(order, lev_names[, 1:nf]), ]
     response <- dat2[, 1]
     nr_hypo <- attr(terms(formula), "factors")
+    outcome_names <- rownames(nr_hypo)[1]  # names of outcome variables
     fac_names <- colnames(nr_hypo)
     fac_names_original <- fac_names
     perm_names <- t(attr(terms(formula), "factors")[-1, ])
@@ -189,10 +195,17 @@ MANOVA <- function(formula, data, subject,
                      .drop = F)
     n <- n$Measure/p
     
-    nested <- grepl(":", formula)
+    # correct formula?
+    if (length(fac_names) != nf && length(fac_names) != nh){
+      stop("Something is wrong with the formula. Please specify all or no interactions.")
+    }
     
-    if (sum(nested) > 0) {
+    nested <- grepl(":", formula)
+    nested2 <- grepl("%in%", formula)
+    
+    if (sum(nested) > 0 || sum(nested2) > 0){
       # nested
+      nest <- TRUE
       
       # if nested factor is named uniquely
       if (nested.levels.unique){
@@ -230,10 +243,28 @@ MANOVA <- function(formula, data, subject,
       hypo_matrices <- HN_MANOVA(fl, p)
     } else {
       # crossed
-      hypo_matrices <- HC_MANOVA(fl, perm_names, fac_names, p)[[1]]
-      fac_names <- HC_MANOVA(fl, perm_names, fac_names, p)[[2]]
+      nest <- FALSE
+      
+      ## adapting formula argument, if interaction term missing
+      
+      if (nrow(perm_names) != nh) {
+        #stop("For crossed designs, an interaction term must be specified in the formula.")
+        form2 <- as.formula(paste(outcome_names, "~", paste(fac_names, collapse = "*")))
+        perm_names2 <- t(attr(terms(form2), "factors")[-1, ])
+        fac_names2 <- attr(terms(form2), "term.labels")
+        hyps <- HC_MANOVA(fl, perm_names2, fac_names2, p, nh)
+        hypo_matrices <- hyps[[1]]
+        fac_names2 <- hyps[[2]]
+        # choose only relevant entries of the hypo matrices
+        indices <- grep(":", fac_names2, invert = T)
+        hypo_matrices <- lapply(indices, function(x) hypo_matrices[[x]])
+        
+      } else {
+        hyps <- HC_MANOVA(fl, perm_names, fac_names, p, nh)
+        hypo_matrices <- hyps[[1]]
+        fac_names <- hyps[[2]]
     }
-    
+    }
     # ---------------------- error detection ------------------------------------
     
     # mixture of nested and crossed designs is not possible
@@ -281,10 +312,12 @@ MANOVA <- function(formula, data, subject,
     Var_out <- results$Cov
     descriptive <- cbind(lev_names, n, mean_out)
     colnames(descriptive) <- c(nadat2, "n", paste(rep("Mean", p), 1:p))
-    
-    # Output ------------------------------------------------------
     colnames(WTS_out) <- cbind ("Test statistic", "df", "p-value")
     colnames(WTPS_out) <- cbind(paste(resampling, "(WTS)"), paste(resampling, "(MATS)"))
+    colnames(MATS_out) <- "Test statistic"
+    
+  }
+  # Output ------------------------------------------------------
     output <- list()
     output$time <- time
     output$input <- input_list
@@ -299,7 +332,19 @@ MANOVA <- function(formula, data, subject,
     output$H <- hypo_matrices
     output$factors <- fac_names
     output$p <- p
+    output$fl <- fl
+    output$BSMeans <- results$BSmeans
+    output$BSVar <- results$BSVar
+    output$levels <- lev_names
+    output$nested <- nest
+  
+  
+  # check for singular covariance matrix
+  test <- try(solve(output$Covariance), silent = TRUE)
+  if(!is.matrix(test)){
+    warning("The covariance matrix is singular. The WTS provides no valid test statistic!")
   }
+  
   class(output) <- "MANOVA"
   return(output)
 }