+++ R/MANOVAwide-function.R
@@ -9,7 +9,8 @@
 #'   contains the matrix of response variables and the right hand side contains the factor 
 #'   variables of interest. An interaction term must be specified.
 #' @param data A data.frame, list or environment containing the variables in 
-#'   \code{formula}. Data must be in wide format.
+#'   \code{formula}. Data must be in wide format. Note: Lines containing missing values will 
+#'   be removed.
 #' @param iter The number of iterations used for calculating the resampled 
 #'   statistic. The default option is 10,000.
 #' @param alpha A number specifying the significance level; the default is 0.05.
@@ -60,6 +61,7 @@ MANOVA.wide <- function(formula, data,
   
   input_list <- list(formula = formula, data = data,
                      iter = iter, alpha = alpha, resampling = resampling)
+  output <- list()
   
   test1 <- hasArg(CPU)
   if(!test1){
@@ -78,15 +80,19 @@ MANOVA.wide <- function(formula, data,
   
   outcome_names <- rownames(nr_hypo)[1]  # names of outcome variables
   # extract names of outcome variables
+  if (grepl("cbind", outcome_names)){
   split1 <- strsplit(outcome_names, "(", fixed = TRUE)[[1]][-1]
   split2 <- strsplit(split1, ")", fixed = TRUE)[[1]]
   split3 <- strsplit(split2, ",")[[1]]
+  } else {
+    split3 <- outcome_names
+  }
   
   EF <- rownames(nr_hypo)[-1]  # names of influencing factors
   nf <- length(EF)
   names(dat) <- c("response", EF)
   #no. dimensions
-  p <- ncol(dat$response)
+  p <- ncol(as.matrix(dat$response))
   fl <- NA
   for (aa in 1:nf) {
     fl[aa] <- nlevels(as.factor(dat[, (aa + 1)]))
@@ -96,61 +102,48 @@ MANOVA.wide <- function(formula, data,
     levels[[jj]] <- levels(as.factor(dat[, (jj + 1)]))
   }
   lev_names <- expand.grid(levels)
+  # number of hypotheses
+  tmp <- 0
+  for (i in 1:nf) {
+    tmp <- c(tmp, choose(nf, i))
+    nh <- sum(tmp)
+  }
+  
+  # MANOVA vs. MANOVAwide?
+  if(sum(grepl("cbind", formula)) == 0 && p != 1){
+    stop("For data in long format, please use function MANOVA()")
+  }
+  
+  # correct formula?
+  if (length(fac_names) != nf && length(fac_names) != nh){
+    stop("Something is wrong with the formula. Please specify all or no interactions in crossed designs.")
+  }
+  # mixture of nested and crossed designs is not possible
+  if (length(fac_names) != nf && 2 %in% nr_hypo) {
+    stop("A model involving both nested and crossed factors is
+           not implemented!")
+  }
   
   if (nf == 1) {
     # one-way layout
+    nest <- FALSE
     dat2 <- dat[order(dat[, 2]), ]
     fac.groups <- dat2[, 2]
-    n.groups <- prod(fl)
-    Y <- split(dat2, fac.groups)
-    n <- sapply(Y, nrow)
-    hypo <- (diag(fl) - matrix(1 / fl, ncol = fl, nrow = fl)) %x% diag(p)
-    
-    WTS_out <- matrix(NA, ncol = 3, nrow = 1)
-    MATS_out <- NA
-    WTPS_out <- rep(NA, 2)
-    quantiles <- matrix(NA, 2, 1)
-    rownames(WTS_out) <- fac_names
-    names(WTPS_out) <- fac_names
-    results <- MANOVA.Stat.wide(Y, n = n, hypo, iter = iter, alpha, resampling, CPU, seed)    
-    WTS_out <- round(results$WTS, dec)
-    MATS_out <- round(results$MATS, dec)
-    WTPS_out <- round(results$WTPS, dec)
-    quantiles <- results$quantiles
-    names(quantiles) <- c("WTS_resampling", "MATS_resampling")
-    mean_out <- matrix(round(results$Mean, dec), ncol = p, byrow = TRUE)
-    Var_out <- results$Cov
-    descriptive <- cbind(lev_names, n, mean_out)
-    colnames(descriptive) <- c(EF, "n", split3)  
-    names(WTS_out) <- cbind ("Test statistic", "df",
-                             "p-value")
-    names(WTPS_out) <- cbind(paste(resampling, "(WTS)"), paste(resampling, "(MATS)"))
-    output <- list()
-    output$input <- input_list
-    output$Descriptive <- descriptive
-    output$Covariance <- Var_out
-    output$WTS <- WTS_out
-    output$MATS <- MATS_out
-    output$resampling <- WTPS_out
-    output$quantile <- quantiles
-    output$nf <- nf
-    output$factors <- fac_names
-    output$H <- hypo
-    output$p <- p
-    output$fl <- fl
-    output$Means <- mean_out
+    hypo_matrices <- list((diag(fl) - matrix(1 / fl, ncol = fl, nrow = fl)) %x% diag(p))
     # end one-way layout ------------------------------------------------------
   } else {
     dat2 <- dat[do.call(order, dat[, 2:(nf + 1)]), ]
     fac.groups <- do.call(list, dat2[, 2:(nf+1)])
-   
+  }
     Y <- split(dat2, fac.groups, lex.order = TRUE)
     n <- sapply(Y, nrow)
     
-    nested <- grepl(":", formula)
+    nested <- grepl(":", formula) 
+    nested2 <- grepl("%in%", formula)
     
-    if (sum(nested) > 0) {
+    if (sum(nested) > 0 || sum(nested2) > 0) {
       # nested
+      nest <- TRUE
       
       # if nested factor is named uniquely
       if (nested.levels.unique){
@@ -160,7 +153,7 @@ MANOVA.wide <- function(formula, data,
         blev <- list()
         lev_names <- list()
         for (ii in 1:length(levels[[1]])) {
-          blev[[ii]] <- levels(as.factor(dat[, 3][dat[, 2] == levels[[1]][ii]]))
+          blev[[ii]] <- droplevels(as.factor(dat[, 3][dat[, 2] == levels[[1]][ii]]))
           lev_names[[ii]] <- rep(levels[[1]][ii], length(blev[[ii]]))
         }
         if (nf == 2) {
@@ -188,12 +181,31 @@ MANOVA.wide <- function(formula, data,
       hypo_matrices <- HN_MANOVA(fl, p)
     } else {
       # crossed
-      hypo_matrices <- HC_MANOVA(fl, perm_names, fac_names, p)[[1]]
+      nest <- FALSE
+      
+      ## adapting formula argument, if interaction term missing
+      if (nrow(perm_names) != nh) {
+        #stop("For crossed designs, an interaction term must be specified in the formula.")
+        form2 <- as.formula(paste(outcome_names, "~", paste(fac_names, collapse = "*")))
+        perm_names2 <- t(attr(terms(form2), "factors")[-1, ])
+        fac_names2 <- attr(terms(form2), "term.labels")
+        hyps <- HC_MANOVA(fl, perm_names2, fac_names2, p, nh)
+        hypo_matrices <- hyps[[1]]
+        fac_names2 <- hyps[[2]]
+        # choose only relevant entries of the hypo matrices
+        indices <- grep(":", fac_names2, invert = T)
+        hypo_matrices <- lapply(indices, function(x) hypo_matrices[[x]])
+        
+      } else if(nf !=1){
+        hyps <- HC_MANOVA(fl, perm_names, fac_names, p, nh)
+        hypo_matrices <- hyps[[1]]
+        fac_names <- hyps[[2]]
+      }
     }
     
+    # correcting for "empty" combinations (if no interaction specified)
     n.groups <- prod(fl)
-    
-    if(length(Y) != n.groups){
+    if(nf != 1 & length(Y) != n.groups){
       index <- NULL
       for(i in 1:length(Y)){
         if(nrow(Y[[i]]) == 0){
@@ -202,15 +214,14 @@ MANOVA.wide <- function(formula, data,
       }
       Y <- Y[-index]
     }
+    Y2 <- lapply(Y, function(x) x$response)
+    if (p==1){
+      Y2 <- lapply(Y2, function(x) as.matrix(x))
+    }
     
     
     # ---------------------- error detection ------------------------------------
     
-    # mixture of nested and crossed designs is not possible
-    if (length(fac_names) != nf && 2 %in% nr_hypo) {
-      stop("A model involving both nested and crossed factors is
-           not implemented!")
-    }
     # only 3-way nested designs are possible
     if (sum(nested) > 0 && nf >= 4) {
       stop("Four- and higher way nested designs are
@@ -236,8 +247,8 @@ MANOVA.wide <- function(formula, data,
     colnames(quantiles) <- c("WTS_resampling", "MATS_resampling")
     # calculate results
     for (i in 1:length(hypo_matrices)) {
-      results <- MANOVA.Stat.wide(Y, n, hypo_matrices[[i]],
-                             iter, alpha, resampling, CPU, seed)
+      results <- MANOVA.Stat.wide(Y2, n, hypo_matrices[[i]],
+                             iter, alpha, resampling, CPU, seed, p)
       WTS_out[i, ] <- round(results$WTS, dec)
       WTPS_out[i, ] <- round(results$WTPS, dec)
       MATS_out[i] <- round(results$MATS, dec)
@@ -247,15 +258,14 @@ MANOVA.wide <- function(formula, data,
     time <- results$time
     mean_out <- matrix(round(results$Mean, dec), ncol = p, byrow = TRUE)
     Var_out <- results$Cov
-    descriptive <- cbind(lev_names, n, mean_out)
+    descriptive <- cbind(unique(lev_names), n, mean_out)
     colnames(descriptive) <- c(EF, "n", split3)
     rownames(descriptive) <- NULL
-    
-    # Output ------------------------------------------------------
     colnames(WTS_out) <- cbind ("Test statistic", "df", "p-value")
     colnames(WTPS_out) <- cbind(paste(resampling, "(WTS)"), paste(resampling, "(MATS)"))
-    output <- list()
-    output$time <- time
+    colnames(MATS_out) <- "Test statistic"
+    
+  # Output ------------------------------------------------------
     output$input <- input_list
     output$Descriptive <- descriptive
     output$Covariance <- Var_out
@@ -268,7 +278,19 @@ MANOVA.wide <- function(formula, data,
     output$H <- hypo_matrices
     output$factors <- fac_names
     output$p <- p
+    output$fl <- fl
+    output$BSMeans <- results$BSmeans
+    output$BSVar <- results$BSVar
+    output$levels <- lev_names
+    output$nested <- nest
+
+
+  # check for singular covariance matrix
+  test <- try(solve(output$Covariance), silent = TRUE)
+  if(!is.matrix(test)){
+    warning("The covariance matrix is singular. The WTS provides no valid test statistic!")
   }
+  
   class(output) <- "MANOVA"
   return(output)
 }