+++ R/confint_MANOVA.R
@@ -8,7 +8,7 @@
 #' @return A \code{confreg} object containing the following components: 
 #'   \item{center}{The center of the confidence ellipsoid.}
 #'   \item{scale}{The scaling factors for the axis of the confidence ellipsoid calculated as \eqn{\sqrt{\lambda*c/N}}, where \eqn{\lambda} are the
-#'   eigenvalues, c denotes the bootstrap quantile and N is the total sample size. See Friedrich and Pauly (2017) for details.}
+#'   eigenvalues, c denotes the bootstrap quantile and N is the total sample size. See Friedrich and Pauly (2018) for details.}
 #'   \item{eigenvectors}{The corresponding eigenvectors, which determine the axes of the ellipsoid.}
 #' 
 #' @examples data(EEG)
@@ -18,11 +18,15 @@
 #' conf.reg(EEG_mod, nullhypo = "sex")
 #' 
 #' 
-#' @references Friedrich, S., and Pauly, M. (2017). MATS: Inference for potentially singular and 
-#'   heteroscedastic MANOVA. arXiv preprint arXiv:1704.03731.
+#' @references Friedrich, S., and Pauly, M. (2018). MATS: Inference for potentially singular and 
+#'   heteroscedastic MANOVA. Journal of Multivariate Analysis, 165, 166-179.
 #'
 #' @export
 conf.reg <- function(object, nullhypo){
+  
+  if(class(object) != "MANOVA"){
+    stop("Object must be of class MANOVA")
+  }
   
   confreg <- list()
 
@@ -36,7 +40,7 @@ conf.reg <- function(object, nullhypo){
   
   if (nf == 1){
   c_star <- object$quantile[2]
-  hypo <- p*object$H
+  hypo <- p*object$H[[1]]
   confreg$nullhypo <- factors
 } else {
   index <- which(factors == nullhypo)