+++ inst/doc/Introduction_to_MANOVA.RM.R
@@ -58,6 +58,39 @@ cr
 plot(cr, col = 2, lty = 2, xlab = "Difference in mortality", ylab ="Difference in water hardness")
 
 ## ------------------------------------------------------------------------
+# pairwise comparison using Tukey contrasts
+simCI(EEG_MANOVA, contrast = "pairwise", type = "Tukey")
+# the same but with Dunnett contrasts using group 2 as baseline
+simCI(EEG_MANOVA, contrast = "pairwise", type = "Dunnett", base = 2)
+# a one-way layout using MANOVA.wide():
+oneway <- MANOVA.wide(cbind(brainrate_temporal, brainrate_central) ~ diagnosis, data = EEGwide,
+                      iter = 1000, CPU = 1)
+# and a user-defined contrast matrix
+H <- as.matrix(cbind(rep(1, 5), -1*Matrix::Diagonal(5)))
+# user-specified comparison
+simCI(oneway, contrast = "user-defined", contmat = H)
+
+## ------------------------------------------------------------------------
+model_sex <- MANOVA.wide(cbind(brainrate_temporal, brainrate_central, brainrate_frontal,
+                            complexity_temporal, complexity_central, complexity_frontal) ~ sex,
+                      data = EEGwide,
+                      iter = 1000, seed = 987, CPU = 1)
+summary(model_sex)
+
+## ------------------------------------------------------------------------
+EEG1 <- MANOVA.wide(brainrate_temporal ~ sex, data = EEGwide, iter = 1000, seed = 987, CPU = 1)
+EEG2 <- MANOVA.wide(brainrate_central ~ sex, data = EEGwide, iter = 1000, seed = 987, CPU = 1)
+EEG3 <- MANOVA.wide(brainrate_frontal ~ sex, data = EEGwide, iter = 1000, seed = 987, CPU = 1)
+EEG4 <- MANOVA.wide(complexity_temporal ~ sex, data = EEGwide, iter = 1000, seed = 987, CPU = 1)
+EEG5 <- MANOVA.wide(complexity_central ~ sex, data = EEGwide, iter = 1000, seed = 987, CPU = 1)
+EEG6 <- MANOVA.wide(complexity_frontal ~ sex, data = EEGwide, iter = 1000, seed = 987, CPU = 1)
+
+## ------------------------------------------------------------------------
+p.adjust(c(EEG1$resampling[, 2], EEG2$resampling[, 2], EEG3$resampling[, 2],
+           EEG4$resampling[, 2], EEG5$resampling[, 2], EEG6$resampling[, 2]),
+         method = "bonferroni")
+
+## ------------------------------------------------------------------------
 if (requireNamespace("GFD", quietly = TRUE)) {
 library(GFD)
 data(curdies)
@@ -65,7 +98,7 @@ set.seed(123)
 curdies$dug2 <- curdies$dugesia + rnorm(36)
 
 # first possibility: MANOVA.wide
-fit1 <- MANOVA.wide(cbind(dugesia, dug2) ~ season + season:site, data = curdies, iter = 100, nested.levels.unique = TRUE, seed = 123, CPU = 1)
+fit1 <- MANOVA.wide(cbind(dugesia, dug2) ~ season + season:site, data = curdies, iter = 1000, nested.levels.unique = TRUE, seed = 123, CPU = 1)
 
 # second possibility: MANOVA (long format)
 dug <- c(curdies$dugesia, curdies$dug2)
@@ -73,7 +106,8 @@ season <- rep(curdies$season, 2)
 site <- rep(curdies$site, 2)
 curd <- data.frame(dug, season, site, subject = rep(1:36, 2))
 
-fit2 <- MANOVA(dug ~ season + season:site, data = curd, subject = "subject", nested.levels.unique = TRUE, seed = 123, iter = 100, CPU = 1)
+fit2 <- MANOVA(dug ~ season + season:site, data = curd, subject = "subject", nested.levels.unique = TRUE, seed = 123, iter = 1000,
+               CPU = 1)
 
 # comparison of results
 summary(fit1)